bugfix/Locale-switcher-select
Replaced the select dropdown in LocaleSwitcherSelect with a custom animated dropdown using framer-motion, improved accessibility, and added outside click handling. Updated the Header logo to be a clickable link to the homepage and adjusted the registration button hover color. Also, improved the BurgerMenu registration button to change text color on hover for better visibility.